src/scraper.py
from playwright.async_api import async_playwright, Browser, Page
from fake_useragent import UserAgent
import asyncio
import random
from datetime import datetime
from loguru import logger
from typing import List, Optional

from models import RealEstateListing
from config import SearchCriteria, Config

class StealthScraper:
    def __init__(self):
        self.ua = UserAgent()
        
    async def _setup_browser(self) -> Browser:
        """Setup a stealth browser instance with randomized properties."""
        playwright = await async_playwright().start()
        browser = await playwright.chromium.launch(
            headless=True,
            args=[
                '--disable-blink-features=AutomationControlled',
                '--disable-dev-shm-usage',
                f'--user-agent={self.ua.random}'
            ]
        )
        return browser

    async def _random_delay(self) -> None:
        """Implement randomized delay between actions."""
        delay = random.uniform(Config.MIN_DELAY, Config.MAX_DELAY)
        await asyncio.sleep(delay)

    async def _simulate_human_behavior(self, page: Page) -> None:
        """Simulate realistic human-like behavior."""
        # Random scroll
        await page.evaluate("""
            window.scrollTo({
                top: Math.random() * document.body.scrollHeight,
                behavior: 'smooth'
            });
        """)
        await self._random_delay()
        
        # Random mouse movements
        for _ in range(random.randint(2, 5)):
            await page.mouse.move(
                random.randint(0, 800),
                random.randint(0, 600)
            )
            await asyncio.sleep(random.uniform(0.1, 0.3))

    async def scrape_kv_ee(self, criteria: SearchCriteria) -> List[RealEstateListing]:
        """
        Scrape real estate listings from kv.ee with stealth measures.
        """
        listings = []
        browser = await self._setup_browser()
        
        try:
            page = await browser.new_page()
            
            # Set custom headers
            await page.set_extra_http_headers({
                'Accept-Language': 'et-EE,et;q=0.9,en-US;q=0.8,en;q=0.7',
                'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8',
                'DNT': '1',
                'Connection': 'keep-alive',
                'Upgrade-Insecure-Requests': '1',
            })

            # Navigate to search page
            search_url = self._build_search_url(criteria)
            await page.goto(search_url)
            await self._simulate_human_behavior(page)

            # Extract listings
            while True:
                # Extract current page listings
                page_listings = await self._extract_page_listings(page)
                listings.extend(page_listings)
                
                # Check for next page
                next_button = await page.query_selector('.pagination__next')
                if not next_button or not await next_button.is_visible():
                    break
                    
                await self._random_delay()
                await next_button.click()
                await page.wait_for_load_state('networkidle')
                await self._simulate_human_behavior(page)

        finally:
            await browser.close()
            
        return listings

    def _build_search_url(self, criteria: SearchCriteria) -> str:
        """Build the search URL based on criteria."""
        # Implementation specific to kv.ee URL structure
        base_url = f"{Config.KV_EE_BASE_URL}/search"
        params = {
            "city": criteria.location,
            "type": criteria.property_type.value,
            "price_min": criteria.min_price,
            "price_max": criteria.max_price,
        }
        # Convert params to URL query string
        return f"{base_url}?{'&'.join(f'{k}={v}' for k, v in params.items() if v is not None)}"

    async def _extract_page_listings(self, page: Page) -> List[RealEstateListing]:
        """Extract listings from the current page."""
        listings = []
        # Implementation specific to kv.ee DOM structure
        # This is a placeholder - actual implementation would need to match kv.ee's HTML structure
        listing_elements = await page.query_selector_all('.real-estate-item')
        
        for element in listing_elements:
            try:
                listing = RealEstateListing(
                    source="kv.ee",
                    listing_url=await element.get_attribute('href'),
                    price=await self._extract_price(element),
                    address=await self._extract_address(element),
                    area_m2=await self._extract_area(element),
                    rooms=await self._extract_rooms(element),
                    listing_time=datetime.utcnow(),  # Placeholder
                    contact_info=await self._extract_contact(element)
                )
                listings.append(listing)
            except Exception as e:
                logger.error(f"Error extracting listing: {e}")
                continue
                
        return listings

    # Helper methods for extracting specific fields
    async def _extract_price(self, element) -> float:
        # Implementation needed
        pass

    async def _extract_address(self, element) -> str:
        # Implementation needed
        pass

    async def _extract_area(self, element) -> float:
        # Implementation needed
        pass

    async def _extract_rooms(self, element) -> Optional[int]:
        # Implementation needed
        pass

    async def _extract_contact(self, element) -> Optional[str]:
        # Implementation needed
        pass